+++ twake/backend/node/src/core/platform/framework/utils/loader.ts
@@ -1,15 +1,28 @@
 import { logger } from "../logger";
 
 export class Loader {
-  constructor(readonly path: string) {}
+  constructor(readonly paths: string[]) {}
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  async load(className: string): Promise<any> {
-    const modulePath = `${this.path}/${className}`;
+  async load(componentName: string): Promise<any> {
+    const modulesPaths = this.paths.map(path => `${path}/${componentName}`);
 
-    logger.debug("Loading %s", modulePath);
-    const clazz = await import(modulePath);
+    logger.debug(`Loading ${componentName} from ${modulesPaths.join(" - ")}`);
 
-    return clazz.default;
+    let classes = await Promise.all(modulesPaths.map(async modulePath => {
+      try {
+        return await import(modulePath);
+      } catch (err) {
+        logger.debug(`${modulePath} not found`);
+      }
+    }));
+
+    classes = classes.filter(Boolean);
+
+    if (!classes || !classes.length) {
+      throw new Error(`Can not find ${componentName} in any given path`);
+    }
+
+    return classes[0].default;
   }
 }
