+++ twake/backend/node/src/core/platform/services/websocket/provider.ts
@@ -1,7 +1,7 @@
 import { EventEmitter } from "events";
 import socketIO from "socket.io";
-import { TwakeServiceProvider } from "../../core/platform/framework";
-import { User } from "../types";
+import { TwakeServiceProvider } from "../../framework";
+import { User } from "../../../../services/types";
 import { WebsocketUserEvent } from "./types";
 
 export default interface WebSocketAPI extends TwakeServiceProvider, EventEmitter {
